requirements.txt
streamlit>=1.18
ultralytics
pyyaml
Pillow
opencv-python-headless
gdown
requests